add AbstractExecutorRepository
