+++ dubbo-common/src/main/java/org/apache/dubbo/common/threadpool/factory/DefaultExecutorRepositoryFactory.java
@@ -34,6 +34,6 @@ public class DefaultExecutorRepositoryFactory implements ExecutorRepositoryFacto
     @Override
     public ExecutorRepository getExecutorRepository(URL url) {
         String port = String.valueOf(url.getUrlAddress());
-        return defaultRepositoryMap.computeIfAbsent(port, key -> new DefaultExecutorRepository(url.getOrDefaultApplicationModel()));
+        return defaultRepositoryMap.computeIfAbsent(port, key -> new DefaultExecutorRepository(url));
     }
 }
